Inline legacy preprocessing helpers
