--- src/gcpvqvae/data/reference/__init__.py
@@ -1,5 +1,0 @@
-"""Reference preprocessing utilities for AlphaFold-style structures."""
-
-from .preprocessing import PreprocessedChain, preprocess_chain, preprocess_structure
-
-__all__ = ["PreprocessedChain", "preprocess_chain", "preprocess_structure"]
